requirements.txt
openai==0.28.0
Flask==3.0.3
googlemaps==4.10.0
pandas==1.5.3
folium==0.17.0
streamlit_folium==0.21.0
python-dotenv==1.0.1
streamlit==1.36.0